Add new login and registration routes
